refactor: split completer.go by responsibility (#24)
Split the 636-line ui/completer.go into four focused files:
- completer.go: core model struct and state management (~260 lines)
- completer_fs.go: DirLister interface, filesystem helpers (~90 lines)
- completer_fuzzy.go: fuzzy matching algorithm and types (~80 lines)
- completer_view.go: rendering, highlighting, box layout (~170 lines)

No behavior changes — pure file-organization refactor.
